fix: fix the failing tests + overall improvements
